Move views files to subfolder
