+++ app/stac_api/views/views_test.py
@@ -3,10 +3,10 @@ import logging
 from rest_framework import generics
 
 from stac_api.models import LandingPage
-from stac_api.views import AssetDetail
-from stac_api.views import CollectionAssetDetail
-from stac_api.views import CollectionDetail
-from stac_api.views import ItemDetail
+from stac_api.views.views import AssetDetail
+from stac_api.views.views import CollectionAssetDetail
+from stac_api.views.views import CollectionDetail
+from stac_api.views.views import ItemDetail
 
 logger = logging.getLogger(__name__)
